# V2.3/code/_openapi_snapshot.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Starriver SuperBrain API",
    "description": "Minimal runnable V2.3 API skeleton",
    "version": "v2.3-preview"
  },
  "paths": {
...
    "/api/v2.3-preview/observability/logs/search": {
      "get": {
        "tags": [
          "observability"
        ],
        "summary": "Search Logs",
        "operationId": "search_logs_api_v2_3_preview_observability_logs_search_get",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "text or tag contains",
              "title": "Q"
            },
            "description": "text or tag contains"
          },
          {
            "name": "level",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "INFO/WARN/ERROR",
              "title": "Level"
            },
            "description": "INFO/WARN/ERROR"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "since_seconds",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer",
                  "minimum": 1
                },
                {
                  "type": "null"
                }
              ],
              "default": 3600,
              "title": "Since Seconds"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "observability"
        ],
        "summary": "Search Logs Post",
        "operationId": "search_logs_post_api_v2_3_preview_observability_logs_search_post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LogSearchRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v2.3-preview/observability/assets": {
      "get": {
        "tags": [
          "observability"
        ],
        "summary": "Get Asset Inventory",
        "operationId": "get_asset_inventory_api_v2_3_preview_observability_assets_get",
        "parameters": [
          {
            "name": "since_seconds",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 600,
              "title": "Since Seconds"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AssetInventoryResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v2.3-preview/observability/migration/plan": {
      "post": {
        "tags": [
          "observability"
        ],
        "summary": "Plan Migration",
        "operationId": "plan_migration_api_v2_3_preview_observability_migration_plan_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MigrationPlanRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MigrationPlanResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v2.3-preview/observability/migration/execute": {
      "post": {
        "tags": [
          "observability"
        ],
        "summary": "Execute Migration",
        "operationId": "execute_migration_api_v2_3_preview_observability_migration_execute_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MigrationPlanRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MigrationPlanResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v2.3-preview/observability/reflection/suggestions": {
      "get": {
        "tags": [
          "observability"
        ],
        "summary": "Reflection Suggestions",
        "operationId": "reflection_suggestions_api_v2_3_preview_observability_reflection_suggestions_get",
        "parameters": [
          {
            "name": "since_seconds",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 900,
              "title": "Since Seconds"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 20,
              "minimum": 1,
              "default": 5,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReflectionResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AssetInventoryResponse": {
        "properties": {
          "modules": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Modules"
          },
          "gauges": {
            "additionalProperties": {
              "type": "number"
            },
            "type": "object",
            "title": "Gauges"
          },
          "counters": {
            "additionalProperties": {
              "type": "integer"
            },
            "type": "object",
            "title": "Counters"
          },
          "timings": {
            "additionalProperties": {
              "additionalProperties": {
                "type": "number"
              },
              "type": "object"
            },
            "type": "object",
            "title": "Timings"
          },
          "experience_rules_total": {
            "type": "integer",
            "title": "Experience Rules Total"
          },
          "experience_candidates_total": {
            "type": "integer",
            "title": "Experience Candidates Total"
          },
          "recent_logs": {
            "type": "integer",
            "title": "Recent Logs"
          },
          "updated_at": {
            "type": "string",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "modules",
          "gauges",
          "counters",
          "timings",
          "experience_rules_total",
          "experience_candidates_total",
          "recent_logs",
          "updated_at"
        ],
        "title": "AssetInventoryResponse"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "LogSearchRequest": {
        "properties": {
          "q": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Q"
          },
          "query": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Query"
          },
          "level": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Level"
          },
          "limit": {
            "type": "integer",
            "title": "Limit",
            "default": 50
          },
          "since_seconds": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Since Seconds",
            "default": 3600
          }
        },
        "type": "object",
        "title": "LogSearchRequest"
      },
...
      "MigrationPlanRequest": {
        "properties": {
          "source_version": {
            "type": "string",
            "title": "Source Version",
            "default": "v2.2"
          },
          "target_version": {
            "type": "string",
            "title": "Target Version",
            "default": "v2.3"
          },
          "dry_run": {
            "type": "boolean",
            "title": "Dry Run",
            "default": true
          },
          "scope": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Scope",
            "default": []
          }
        },
        "type": "object",
        "title": "MigrationPlanRequest"
      },
...
      "ReflectionResponse": {
        "properties": {
          "count": {
            "type": "integer",
            "title": "Count"
          },
          "suggestions": {
            "items": {
              "$ref": "#/components/schemas/ReflectionSuggestion"
            },
            "type": "array",
            "title": "Suggestions"
          },
          "updated_at": {
            "type": "string",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "count",
          "suggestions",
          "updated_at"
        ],
        "title": "ReflectionResponse"
      },
...
    }
  }
}